+++ src/main/java/com/eazy/brush/dao/entity/TaskSub.java
@@ -11,7 +11,7 @@ import lombok.Data;
 public class TaskSub {
     private int id;
     private int taskId;              //任务id
-    private int perTime;             //时间粒度
+    private long perTime;             //时间粒度
     private int actionId;            //动作组id
     private int deviceInfoId;        //设备信息
     private int cardInfoId;          //手机卡信息
